changing argument name and renaming files
